feat(fire-incident): add incident boundary utils and optimized country data
